Change type to ssize_t for all index and size variables

+++ skimage/_shared/interpolation.pxd
@@ -1,27 +1,27 @@
 
-cdef double nearest_neighbour_interpolation(double* image, int rows,
-                                            int cols, double r,
+cdef double nearest_neighbour_interpolation(double* image, ssize_t rows,
+                                            ssize_t cols, double r,
                                             double c, char mode,
                                             double cval)
 
-cdef double bilinear_interpolation(double* image, int rows, int cols,
+cdef double bilinear_interpolation(double* image, ssize_t rows, ssize_t cols,
                                    double r, double c, char mode,
                                    double cval)
 
 cdef double quadratic_interpolation(double x, double[3] f)
-cdef double biquadratic_interpolation(double* image, int rows, int cols,
+cdef double biquadratic_interpolation(double* image, ssize_t rows, ssize_t cols,
                                       double r, double c, char mode,
                                       double cval)
 
 cdef double cubic_interpolation(double x, double[4] f)
-cdef double bicubic_interpolation(double* image, int rows, int cols,
+cdef double bicubic_interpolation(double* image, ssize_t rows, ssize_t cols,
                                   double r, double c, char mode,
                                   double cval)
 
-cdef double get_pixel2d(double* image, int rows, int cols, int r, int c,
-                        char mode, double cval)
+cdef double get_pixel2d(double* image, ssize_t rows, ssize_t cols, ssize_t r,
+                        ssize_t c, char mode, double cval)
 
-cdef double get_pixel3d(double* image, int rows, int cols, int dims, int r,
-                        int c, int d, char mode, double cval)
+cdef double get_pixel3d(double* image, ssize_t rows, ssize_t cols, ssize_t dims,
+                        ssize_t r, ssize_t c, ssize_t d, char mode, double cval)
 
-cdef int coord_map(int dim, int coord, char mode)
+cdef ssize_t coord_map(ssize_t dim, ssize_t coord, char mode)
