+++ skimage/_shared/transform.pyx
@@ -6,7 +6,7 @@ cimport numpy as cnp
 
 
 cdef float integrate(cnp.ndarray[float, ndim=2,  mode="c"] sat,
-                     int r0, int c0, int r1, int c1):
+                     ssize_t r0, ssize_t c0, ssize_t r1, ssize_t c1):
     """
     Using a summed area table / integral image, calculate the sum
     over a given window.
